Judicial Operation Management Systems
The JOMS is an integrated e-filing and e-workflow case management system for the criminal courts. It tracks criminal and court cases from registration to disposal, enabling enforcement agencies to commence criminal prosecutions and other criminal matters electronically at any time of the day, without making a trip to the courthouse. In addition, it enables accurate, complete and timely sharing, retrieval and exchange of relevant information on case status, outcomes and court orders, thereby facilitating greater accessibility, efficiency, transparency and effectiveness for the stakeholders in the criminal justice system.